+++ synthesis/population/spatial/primary/locations.py
@@ -6,10 +6,14 @@ def configure(context):
     context.stage("synthesis.population.spatial.primary.candidates")
     context.stage("synthesis.population.spatial.commute_distance")
     context.stage("synthesis.population.spatial.home.locations")
-    context.stage("synthesis.destinations")
+    context.stage("synthesis.locations.work")
+    context.stage("synthesis.locations.education")
 
-def define_distance_ordering(df_persons, df_candidates):
+def define_distance_ordering(df_persons, df_candidates, progress):
     indices = []
+
+    f_available = np.ones((len(df_candidates),), dtype = bool)
+    costs = np.ones((len(df_candidates),)) * np.inf
 
     commute_coordinates = np.vstack([
         df_candidates["geometry"].x.values,
@@ -17,14 +21,22 @@ def define_distance_ordering(df_persons, df_candidates):
     ]).T
 
     for home_coordinate, commute_distance in zip(df_persons["home_location"], df_persons["commute_distance"]):
-        distances = np.sqrt(np.sum((commute_coordinates - home_coordinate)**2, axis = 1))
-        costs = np.abs(distances - commute_distance)
-        costs[indices] = np.inf
-        indices.append(np.argmin(costs))
+        distances = np.sqrt(np.sum((commute_coordinates[f_available] - home_coordinate)**2, axis = 1))
+        costs[f_available] = np.abs(distances - commute_distance)
+
+        selected_index = np.argmin(costs)
+        indices.append(selected_index)
+        f_available[selected_index] = False
+        costs[selected_index] = np.inf
+
+        progress.update()
+
+    assert len(set(indices)) == len(df_candidates)
 
     return indices
 
-def define_random_ordering(df_persons, df_candidates):
+def define_random_ordering(df_persons, df_candidates, progress):
+    progress.update(len(df_candidates))
     return np.arange(len(df_candidates))
 
 define_ordering = define_distance_ordering
@@ -37,25 +49,25 @@ def process_municipality(context, origin_id):
     df_persons = df_persons[df_persons["commune_id"] == origin_id][[
         "person_id", "home_location", "commute_distance"
     ]].copy()
-    df_candidates = df_candidates[df_candidates["origin_commune_id"] == origin_id]
+    df_candidates = df_candidates[df_candidates["origin_id"] == origin_id]
 
     # From previous step, this should be equal!
     assert len(df_persons) == len(df_candidates)
 
-    indices = define_ordering(df_persons, df_candidates)
+    indices = define_ordering(df_persons, df_candidates, context.progress)
     df_candidates = df_candidates.iloc[indices]
 
     df_candidates["person_id"] = df_persons["person_id"].values
-    df_candidates = df_candidates.rename(columns = dict(destination_commune_id = "commune_id"))
+    df_candidates = df_candidates.rename(columns = dict(destination_id = "commune_id"))
 
-    return df_candidates[["person_id", "commune_id", "destination_id", "geometry"]]
+    return df_candidates[["person_id", "commune_id", "location_id", "geometry"]]
 
 def process(context, purpose, df_persons, df_candidates):
-    unique_ids = df_candidates["origin_commune_id"].unique()
+    unique_ids = df_candidates["origin_id"].unique()
 
     df_result = []
 
-    with context.progress(label = "Distributing %s destinations" % purpose, total = len(unique_ids)) as progress:
+    with context.progress(label = "Distributing %s destinations" % purpose, total = len(df_persons)) as progress:
         with context.parallel(dict(df_persons = df_persons, df_candidates = df_candidates)) as parallel:
             for df_partial in parallel.imap_unordered(process_municipality, unique_ids):
                 df_result.append(df_partial)
@@ -88,14 +100,14 @@ def execute(context):
     df_education = pd.merge(df_education, df_commute_distance["education"], how = "left", on = "person_id")
 
     # Attach geometry
-    df_destinations = context.stage("synthesis.destinations")[["destination_id", "geometry"]]
-
+    df_locations = context.stage("synthesis.locations.work")[["location_id", "geometry"]]
     df_work_candidates = data["work_candidates"]
-    df_work_candidates = pd.merge(df_work_candidates, df_destinations, how = "left", on = "destination_id")
+    df_work_candidates = pd.merge(df_work_candidates, df_locations, how = "left", on = "location_id")
     df_work_candidates = gpd.GeoDataFrame(df_work_candidates)
 
+    df_locations = context.stage("synthesis.locations.education")[["location_id", "geometry"]]
     df_education_candidates = data["education_candidates"]
-    df_education_candidates = pd.merge(df_education_candidates, df_destinations, how = "left", on = "destination_id")
+    df_education_candidates = pd.merge(df_education_candidates, df_locations, how = "left", on = "location_id")
     df_education_candidates = gpd.GeoDataFrame(df_education_candidates)
 
     # Assign destinations